requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
PyPDF2==3.0.1
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
Wikipedia-API==0.8.1
